Moved backend/ and frontend/ a level up
