+++ src/tasks/start/backend/apm.ts
@@ -1,4 +1,7 @@
 import { hash as namehash } from 'eth-ens-namehash'
+import Web3 from 'web3'
+import { TruffleEnvironmentArtifacts } from '@nomiclabs/buidler-truffle5/src/artifacts'
+import { getLog } from '../../../utils/getLog'
 import {
   DAOFactoryInstance,
   APMRegistryInstance,
@@ -11,9 +14,6 @@ import {
   ENSSubdomainRegistrarInstance,
   APMRegistryFactoryInstance
 } from '~/typechain'
-import Web3 from 'web3'
-import { TruffleEnvironmentArtifacts } from '@nomiclabs/buidler-truffle5/src/artifacts'
-import { getLog } from '../../../../utils/getLog'
 
 const ZERO_ADDR = '0x0000000000000000000000000000000000000000'
