+++ src/tasks/start/frontend/client.ts
@@ -2,8 +2,8 @@ import path from 'path'
 import fs from 'fs'
 import fsExtra from 'fs-extra'
 import os from 'os'
-import { execaLogTo } from '../execa'
-import { logFront } from '../logger'
+import { execaLogTo } from '../utils/execa'
+import { logFront } from '../utils/logger'
 import liveServer from 'live-server'
 import open from 'open'
