+++ test/tasks/start/backend/ganache.test.ts
@@ -1,15 +1,12 @@
 import { assert } from 'chai'
-import {
-  startGanache,
-  stopGanache
-} from '~/src/tasks/start/utils/backend/ganache'
+import { startGanache, stopGanache } from '~/src/tasks/start/backend/ganache'
+import http from 'http'
+import { testnetPort } from '~/src/params'
+import tcpPortUsed from 'tcp-port-used'
 import {
   useDefaultEnvironment,
   useEnvironment
 } from '~/test/test-helpers/useEnvironment'
-import http from 'http'
-import { testnetPort } from '~/src/params'
-import tcpPortUsed from 'tcp-port-used'
 
 describe('ganache.ts', async function() {
   describe('when using the buidlerevm network', async function() {
